Setup paging
- Implemented mapping to physical pages to virtal
- Added serial output: serial.c::serial_write(char)
- Parsing of multiboot2 tags to get memory map
- Setup basic printk (Uses serial device for output)
- Started LIBC\K

+++ src/kernel/include/stddef.h
@@ -1,6 +1,8 @@
 #pragma once
 
-typedef unsigned long long size_t;
+#include <stdint.h>
+
+typedef uint64_t size_t;
 
 #define NULL 0x00
